iscsi: change the format of the 'luns' parameter
Previously, the parameter was a list of luns in <name>=<id> format, leaving no
room to specify the HBA type (iblock was assumed) and, in my opinion, in the
reverse order to what makes sense. This patch changes the format of this
parameter to be <id>:<hba>/<name>, e.g. "0:iblock/test1".

Signed-off-by: Chris Boot <crb@tiger-computing.co.uk>