backstore: Use an ALUA port group per instance of the RA
When using the default port group (ID 0) on two different nodes in an HA setup,
initiators are confused by the port group being the same on both nodes. For
example multipathd with the ALUA hardware handler confuses the ports on both
targets with each other, and the port state bounces constantly when in failover
state, and nothing works. Using a different port group ID on both sides should
cure the problem.

This commit creates a new port group on the LUN for each instance. The name is
the hostname of the machine the RA is running on, and the ID is 16 + the
instance ID from pacemaker.

Signed-off-by: Chris Boot <crb@tiger-computing.co.uk>