backstore: Add alua_hosts parameter to keep consistent ALUA port group IDs
Previously, we generated ALUA port group ID numbers based on Pacemaker's
instance ID number, but it turns out this is volatile and nodes can switch ID
numbers. tcm_node checks that alua_tg_pt_gp_id numbers are kept consistent for
the same name, so this prevented the node coming up.

Fix this by having the resource pass a fixed list of hostnames involved in
exporting this resource, and generate the ID number based on the index into this
list.

Signed-off-by: Chris Boot <crb@tiger-computing.co.uk>